% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{load_from_usfl_path}
\alias{load_from_usfl_path}
\title{Perform USFL API Request}
\usage{
load_from_usfl_path(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[httr2:req_url]{httr2::req_url_path_append}}
  \describe{
    \item{\code{req}}{A \link[httr2]{request}.}
  }}
}
\value{
API response
}
\description{
Perform USFL API Request
}
