% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_scores.R
\name{load_usfl_scores}
\alias{load_usfl_scores}
\title{Load Scores of Multiple USFL Weeks}
\usage{
load_usfl_scores(weeks = 1:10)
}
\arguments{
\item{weeks}{Weeks to load and combine}
}
\value{
A dataframe
}
\description{
This is a wrapper around \code{\link[=load_usfl_scores_in_week]{load_usfl_scores_in_week()}}
}
