% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_usfl_teams.R
\name{load_usfl_teams}
\alias{load_usfl_teams}
\title{Load USFL Teams File from Anthony Reinhard's Github}
\usage{
load_usfl_teams()
}
\value{
A dataframe
}
\description{
Load USFL Teams File from Anthony Reinhard's Github
}
\examples{
load_usfl_teams()
}
