% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusfl-package.R
\docType{package}
\name{rusfl-package}
\alias{rusfl}
\alias{rusfl-package}
\title{rusfl: Load and Parse USFL Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
