% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usfl_abbreviation_from_full_name}
\alias{usfl_abbreviation_from_full_name}
\title{Get valid USFL Team Abbreviations from Full Names}
\usage{
usfl_abbreviation_from_full_name(full_name = NULL)
}
\arguments{
\item{full_name}{A valid USFL full team name for which to output the abbreviation.
The default \code{NULL} returns all abbreviations.}
}
\value{
A named vector.
}
\description{
Get valid USFL Team Abbreviations from Full Names
}
