% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_game_data.R
\name{usfl_load_game_data}
\alias{usfl_load_game_data}
\title{Load Raw Game Data form USFL API}
\usage{
usfl_load_game_data(usfl_game_id)
}
\arguments{
\item{usfl_game_id}{USFL game id, e.g. "usfl2"}
}
\value{
list of game data
}
\description{
Load Raw Game Data form USFL API
}
