% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_roster.R
\name{usfl_load_roster_team}
\alias{usfl_load_roster_team}
\title{Load and Parse USFL Roster for a Given Team}
\usage{
usfl_load_roster_team(team_abbr)
}
\arguments{
\item{team_abbr}{Valid USFL team abbreviation}
}
\value{
A tibble with team roster
}
\description{
Load and Parse USFL Roster for a Given Team
}
