% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_roster.R
\name{usfl_load_rosters}
\alias{usfl_load_rosters}
\title{Load Rosters of Multiple USFL Teams}
\usage{
usfl_load_rosters(teams = NULL)
}
\arguments{
\item{teams}{Valid USFL team abbreviations}
}
\value{
A tibble of rosters of the given teams
}
\description{
Load Rosters of Multiple USFL Teams
}
