% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_scores.R
\name{usfl_load_scores}
\alias{usfl_load_scores}
\title{Load Scores of Multiple USFL Weeks}
\usage{
usfl_load_scores(weeks = 1:10)
}
\arguments{
\item{weeks}{Weeks to load and combine}
}
\value{
A dataframe
}
\description{
This is a wrapper around \code{\link[=usfl_load_scores_in_week]{usfl_load_scores_in_week()}}
}
