% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_scores.R
\name{usfl_load_scores_in_week}
\alias{usfl_load_scores_in_week}
\title{Loads and Parses USFL Scores for a Given Week}
\usage{
usfl_load_scores_in_week(week)
}
\arguments{
\item{week}{week to load}
}
\value{
A dataframe
}
\description{
Loads and Parses USFL Scores for a Given Week
}
