% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_teams.R
\name{usfl_load_teams}
\alias{usfl_load_teams}
\title{Load USFL Teams File from Anthony Reinhard's Github}
\usage{
usfl_load_teams()
}
\value{
A dataframe
}
\description{
Load USFL Teams File from Anthony Reinhard's Github
}
\examples{
usfl_load_teams()
}
