% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfl_game_data.R
\name{usfl_parse_pbp}
\alias{usfl_parse_pbp}
\title{Extract and Parse PBP Date from Raw USFL Game Data}
\usage{
usfl_parse_pbp(raw_game_data)
}
\arguments{
\item{raw_game_data}{Raw data loaded with \code{\link[=usfl_load_game_data]{usfl_load_game_data()}}}
}
\value{
A tibble with pbp
}
\description{
Extract and Parse PBP Date from Raw USFL Game Data
}
