% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usfl_set_api_key}
\alias{usfl_set_api_key}
\title{Save USFL API Key in User Environment}
\usage{
usfl_set_api_key(key = NULL)
}
\arguments{
\item{key}{}
}
\value{
NULL invisibly
}
\description{
Save USFL API Key in User Environment
}
