% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usfl_team_abbreviations}
\alias{usfl_team_abbreviations}
\title{Get valid USFL Team Abbreviations and Corresponding API IDs}
\usage{
usfl_team_abbreviations(abbr = NULL)
}
\arguments{
\item{abbr}{A valid abbreviation for which to output the ID and name.
The default \code{NULL} returns all IDs and names.}
}
\value{
A named vector.
}
\description{
Get valid USFL Team Abbreviations and Corresponding API IDs
}
\examples{
usfl_team_abbreviations()
usfl_team_abbreviations(c("BHAM", "HOU"))
}
